encoder/requirements.txt
ffmpeg-python
pika
google-cloud-storage